+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Added option to Controller to permit caching unsafe HTTP methods.
+- Added option to Controller to permit caching unsafe HTTP methods. (#159)
+- Make `FileStorage` to check staleness of all cache files with set interval. (#169)
 - Support AWS S3 storages. (#164)
 - Move `typing_extensions` from requirements.txt to pyproject.toml. (#161)